10.03

+++ src/Controller/AlbumController.php
@@ -7,25 +7,19 @@ use App\Entity\Piste;
 use App\Form\AlbumType;
 use App\Repository\UserRepository;
 use App\Repository\AlbumRepository;
-use App\Repository\PisteRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
-use Symfony\Component\Security\Http\Attribute\IsGranted;
 use Symfony\Component\DependencyInjection\Attribute\Autowire;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 
-#[Route('/artiste', name: 'artiste_')]
 class AlbumController extends AbstractController
 {
 
     // Methode qui renvoie la vue de tout les albums.
-    // Prend en argument AlbumRepository, cette classe permet de récuperer les données de l'entité Album
-    // Response : renvoie une réponse HTTP
-
-    #[Route('/', name: 'all_album')]
+    #[Route('/album', name: 'app_album')]
     public function index(AlbumRepository $albumRepository): Response
     {
         // Albums stocke les données de AlbumRepository
@@ -41,17 +35,120 @@ class AlbumController extends AbstractController
 
 
     // Méthode qui permet d'ajouter ou d'éditer un album
-    // la route edit prend l'id de l'album 
-    #[Route('/{id}/edit', name: 'edit_album')]
-    #[Route('/new', name: 'new_album')]
-    public function newEdit(Album $album = null, Request $request, EntityManagerInterface $entityManager, #[Autowire('%photo_dir%')]string $photoDir,  #[Autowire('%audio_dir%')]string $audioDir): Response
+    #[Route('/artiste/{slug}/album/{albumId}/edit', name: 'album_edit')]
+    public function edit(
+        $slug, 
+        $albumId, 
+        Request $request,
+        AlbumRepository $albumRepository, 
+        EntityManagerInterface $entityManager, 
+        #[Autowire('%photo_dir%')]string $photoDir,  
+        #[Autowire('%audio_dir%')]string $audioDir
+    ): Response
+    {
+
+        // Si l'utilisateur à le rôle artiste
+        $this->denyAccessUnlessGranted('ROLE_ARTISTE');
+
+        $album = $albumRepository->find($albumId);
+
+        // form stocke le formulaire crée en appellant la méthode createForm, cette méthode prend en argument la classe
+        // AlbumType ainsi que l'objet Album
+        $form = $this->createForm(AlbumType::class, $album);
+
+        // Form récupère les données de la requète en prenant en argument l'objet requete
+        $form->handleRequest($request);
+
+        // Si le formulaire a été soumis et est valide
+        if ($form->isSubmitted() && $form->isValid()) {
+
+            // On stocke les données recupérées dans la variable $album
+            $album = $form->getData();
+
+            // Si dans le tableau $form la propriété 'avatar' existe. On stocke la photo dans la variable $photo
+            // et on rentre dans la condition
+            $photo = $form['photo']->getData();
+
+            
+            // $audio = $form->get('pistes')->getData();
+            // $brochureFile = $form->get('brochure')->getData();
+
+            if ($photo) {
+                // uniqid génère un identifiant unique 
+                // photo récupère l'extension en utilisant la méthode guessExtension de la classe File utilisées dans le formulaire RoleArtisteType
+                // fileName stocke la concatenation de uniqid et de photo
+                $fileName = uniqid().'.'.$photo->guessExtension();
+
+                // photo déplace déplace $fileName dans le dossier précisé dans la variable $photoDir.
+                // move prend en premier argument, le dossier de redirection et le fichier
+                $photo->move($photoDir, $fileName);
+
+                // $album modifie le nom du fichier contenu dans l'oject album
+                $album->setImageAlbum($fileName);
+
+            }
+
+            $audioFiles = $form->get('pistes')->getData();
+
+            foreach ($audioFiles as $audioFile) {
+                // Générez un nom de fichier unique pour chaque piste audio
+                $audioFileName = uniqid().'.'.$audioFile->guessExtension();
+            
+                // Déplacez le fichier audio vers le répertoire approprié
+                $audioFile->move($audioDir, $audioFileName);
+            
+                // Créez une nouvelle instance de piste audio et associez-la à l'album
+                $piste = new Piste();
+                $piste->setAudio($audioFileName); // Supposez que setSon est la méthode pour stocker le nom du fichier audio
+            
+                // Ajoutez la piste à l'album
+                $album->addPiste($piste);
+            
+                // Persistez la piste
+                $entityManager->persist($piste);
+            }
+
+            // On persiste l'album
+            $entityManager->persist($album);
+
+            // On execute la requete en base de donnée après avoir préparé la requète via la méthode persist 
+            $entityManager->flush();
+
+            // Redirige vers la vue 'artiste_detail_album' avec en paramètre l'id de l'album récupéré avec la méthode getId de
+            // l'object album
+            return $this->redirectToRoute('album_detail', ['slug' => $slug, 'albumId' => $album->getId()]);
+
+        }
+
+        // Retourne la vue de formulaire. Prend en argument la vue, la variable form du formulaire, et la variable edit pour la fo
+        // fonction édition 
+        return $this->render('artiste_page/album/new_edit.html.twig', [
+
+            // createView : méthode de la classe formInterface qui crée la vue du formulaire
+            'form' => $form->createView(),
+            'edit' => true,
+            // edit : variable qui affiche l'album à éditer.
+            // Récupère l'Id de l'album via la méthode getId de la classe Album
+            'albumId' => $album->getId(),
+
+        ]);
+    }
+
+    #[Route('/artiste/{slug}/album/new', name: 'album_new')]
+    public function new(
+        $slug, 
+        Album $album, 
+        Request $request, 
+        EntityManagerInterface $entityManager, 
+        #[Autowire('%photo_dir%')]string $photoDir,  
+        #[Autowire('%audio_dir%')]string $audioDir
+    ): Response
     {
 
         // Si l'utilisateur à le rôle artiste
         $this->denyAccessUnlessGranted('ROLE_ARTISTE');
         
         // Si l'album n'existe pas
-        if (!$album) {
 
             // Un nouvel object Album est crée
             $album = new Album();
@@ -61,8 +158,6 @@ class AlbumController extends AbstractController
 
             // La méthode addPiste ajoute un nouvel objet Piste 
             $album->addPiste(new Piste());
-        }
-
 
 
         // form stocke le formulaire crée en appellant la méthode createForm, cette méthode prend en argument la classe
@@ -85,7 +180,7 @@ class AlbumController extends AbstractController
             
             // $audio = $form->get('pistes')->getData();
             // $brochureFile = $form->get('brochure')->getData();
-            dd($form->get('audio')->getData());
+            $pistes = $form->get('pistes')->getData();
 
             if ($photo) {
                 // uniqid génère un identifiant unique 
@@ -104,25 +199,18 @@ class AlbumController extends AbstractController
             }
 
 
-            if($audio){
+            if($pistes){
 
-                $fileName = uniqid().'.'.$audio->guessExtension();
+                foreach ($pistes as $piste) {
+                    $fileName = uniqid().'.'.$piste->getClientOriginalName();
+                    $piste->setAlbum($fileName);
+                    dd($piste);
 
-                $audio->move($audioDir, $fileName);
+                    $album->addPiste($piste);
 
-                $album->addPiste($audio);
-
-            //     // $audio->move($this->getParameter())
-            //     // foreach ($audio as $piste) {
-            //     //     $fileName = uniqid().'.'.$piste->getClientOriginalName();
-            //     //     $piste->setAlbum($fileName);
-            //     //     dd($piste);
-
-            //     //     $album->addPiste($piste);
-
-            //     //     $entityManager->persist($piste);
-            //     //     $entityManager->flush();
-            //     // }
+                    $entityManager->persist($piste);
+                    $entityManager->flush();
+                }
             }
             
 
@@ -134,59 +222,55 @@ class AlbumController extends AbstractController
 
             // Redirige vers la vue 'artiste_detail_album' avec en paramètre l'id de l'album récupéré avec la méthode getId de
             // l'object album
-            return $this->redirectToRoute('artiste_detail_album', ['idAlbum' => $album->getId()]);
+            return $this->redirectToRoute('album_detail', ['slug' => $slug, 'albumId' => $album->getId()]);
 
         }
 
         // Retourne la vue de formulaire. Prend en argument la vue, la variable form du formulaire, et la variable edit pour la fo
         // fonction édition 
-        return $this->render('artiste_page/album/new_edit_album.html.twig', [
+        return $this->render('artiste_page/album/new_edit.html.twig', [
 
             // createView : méthode de la classe formInterface qui crée la vue du formulaire
             'form' => $form->createView(),
-
-            // edit : variable qui affiche l'album à éditer.
+            'edit' => false,
             // Récupère l'Id de l'album via la méthode getId de la classe Album
-            'edit' => $album->getId(),
+            'albumId' => $album->getId(),
 
         ]);
     }
 
     // Fonction pour supprimer un album
-    // route qui prend en paramètre l'id de l'album 
-    // Méthode qui prend en argument l'objet Album, la classe EnityManagerInterface
-    // Response : renvoie une réponse HTTP
-    #[Route('/{id}/delete', name: 'delete_album')]
-    public function deleteAlbum(Album $album, EntityManagerInterface $entityManager): Response
+    #[Route('/artiste/{slug}/album/{albumId}/delete', name: 'album_delete')]
+    public function delete($slug, $albumId, AlbumRepository $albumRepository,EntityManagerInterface $entityManager): Response
     {
 
+        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY');
+
         // L'objet $album sera supprimé
+        $album = $albumRepository->find($albumId);
+
         $entityManager->remove($album);
 
         // Flush execute la requete
         $entityManager->flush();
 
         // Redirection vers la page app_album après effacement de la bdd 
-        return $this->redirectToRoute('app_album');
+        return $this->redirectToRoute('albums_per_artiste', ['slug' => $slug]);
     }
 
     // Méthode qui affiche tout les albums qu'un artiste
-    // Prend en argument l'id de l'utilisateur
-    // AlbumRepository récupère les albums
-    // UserRepository récupère les utilisteurs
-    // Response : renvoie une réponse HTTP
-    #[Route('/album/{id}', name: 'all_album_per_artiste')]
-    public function allAlbumPerArtiste(AlbumRepository $albumRepository, UserRepository $userRepository, $id): Response
+    #[Route('/artiste/{slug}/album/', name: 'albums_per_artiste')]
+    public function albumsPerArtiste(
+        AlbumRepository $albumRepository, 
+        UserRepository $userRepository, 
+        $slug
+    ): Response
     {
 
         // On récupère l'utilisateur en appellant la méthode find de UserRepository
         // Cette méthode prend en argument l'id de la route mise en argument de la fonction allAlbumPerArtiste
-        $user = $userRepository->find($id);
+        $user = $userRepository->findOneBy(['slug' => $slug]);
 
-        // On récupère les albums de l'utilisateur en utilisant la méthode findBy de AlbumRepository
-        // Cette méthode prend en premier argument le user
-        // La méthode findBy prend en 1er argument un critère (équivalent à WHERE en SQL), ensuite un tableau (ORDER BY, GROUP BY en SQL), une limit ( par exemple 'limit:2') et un offset (spécifie le décalage à partir de combien les résultats doivent être
-        // retournés ( offset:10), par défaut sa valeur est $offset = null))
         $albums = $albumRepository->findBy(["user" => $user]);
 
 
@@ -198,25 +282,22 @@ class AlbumController extends AbstractController
             'user' => $user,
         ]);
     }
-
-    // La méthode detailAlbum prend en argument la classe AlbumRepository ainsi que l'id de l'album 
-    // Reponse renvoie une réponse HTTP
     
     // Détails d'un album d'un artiste
-    #[Route('/detail/{idAlbum}', name: 'detail_album')]
-    public function detailAlbum($idAlbum, AlbumRepository $albumRepository): Response
+    #[Route('/artiste/{slug}/album/detail/{albumId}', name: 'album_detail')]
+    public function detail($slug, $albumId, AlbumRepository $albumRepository, UserRepository $userRepository): Response
     {
 
         // On récupère l'album via la méthode findOneBy de AlbumRepository
         // On prend en argument l'id de l'album
-        $album = $albumRepository->findOneBy(['id' => $idAlbum]);
+        $album = $albumRepository->find($albumId);
 
         // On stocke l'utilisateur dans $user en récupèrant l'user de l'album avec la méthode getUser de la classe Album
-        $user = $album->getUser();
+        $user = $userRepository->findOneBy(['slug' => $slug]);
 
         // Renvoie la vue du detail de l'album en utilisant la méthode rendre
         // Cette méthode prend en argument la vue et un tableau de paramètre content l'objet Album et User
-        return $this->render('artiste_page/album/detail_album.html.twig', [
+        return $this->render('artiste_page/album/detail.html.twig', [
             'album' => $album,
             'user' => $user,
         ]);
